chore: ajust folder structure
