+++ src/connect.ts
@@ -1,6 +1,5 @@
 import { MapStateToProps, connect as reactConnect } from 'react-redux'
 import { bindActionCreators, Dispatch } from 'redux'
-import { Action } from 'redux-actions'
 
 import { EffectModule } from './Module'
 
@@ -10,8 +9,4 @@ export const connect = <P, S>(mapStateToProps: MapStateToProps<P, S>, effectModu
   return (comp: any) => reactConnect(mapStateToProps, (componentDispatch: Dispatch<P>) => bindActionCreators(
     dispatchMap, componentDispatch
   ))(comp) as any
-}
-
-export type ModuleActionProps <S, T extends EffectModule<S>> = {
-  [key in keyof T]: (...args: any[]) => Action<any>
 }
