+++ src/symbol.ts
@@ -1,0 +1,6 @@
+export const symbolNamespace = Symbol('namespace')
+export const symbolDispatch = Symbol('dispatch')
+export const symbolReducerMap = Symbol('reducerMap')
+export const symbolEpics = Symbol('actionQueue')
+export const symbolAction = Symbol('action')
+export const symbolNotTrasfer = Symbol('notTrasnfer')
